Refactored parallel interface behind build tags
